LayoutTests/webarchive/ignore-noscript-if-scripting-enabled.html

<html>
<head>
<script>
    if (window.layoutTestController)
        layoutTestController.dumpDOMAsWebArchive();
</script>
</head>
<body>
This should appear in the webarchive.
<noscript>
Assuming scripting was enabled earlier, this should *not* appear in the webarchive.
</noscript>
</body>
</html>
