LayoutTests/webarchive/ignore-noscript-if-scripting-enabled-expected.txt

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
<plist version="1.0">
<dict>
	<key>WebMainResource</key>
	<dict>
		<key>WebResourceData</key>
		<string>&lt;html&gt;&lt;head&gt;
&lt;script&gt;
    if (window.layoutTestController)
        layoutTestController.dumpDOMAsWebArchive();
&lt;/script&gt;
&lt;/head&gt;
&lt;body&gt;
This should appear in the webarchive.



&lt;/body&gt;&lt;/html&gt;</string>
		<key>WebResourceFrameName</key>
		<string></string>
		<key>WebResourceMIMEType</key>
		<string>text/html</string>
		<key>WebResourceTextEncodingName</key>
		<string>UTF-8</string>
		<key>WebResourceURL</key>
		<string>file:///LayoutTests/webarchive/ignore-noscript-if-scripting-enabled.html</string>
	</dict>
</dict>
</plist>
